screencapture in 16-bit
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@4053 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ OrthogonalMPRController.h
@@ -24,7 +24,7 @@
 	OrthogonalReslice			*reslicer;
 	float						sign;
 	
-	long						originalCrossPositionX, originalCrossPositionY, xReslicedCrossPositionX, xReslicedCrossPositionY, yReslicedCrossPositionX, yReslicedCrossPositionY;
+	float						originalCrossPositionX, originalCrossPositionY, xReslicedCrossPositionX, xReslicedCrossPositionY, yReslicedCrossPositionX, yReslicedCrossPositionY;
 	long						orientationVector;
 	
 	IBOutlet OrthogonalMPRView	*originalView, *xReslicedView, *yReslicedView;
